+++ src/main.js
@@ -5,21 +5,17 @@ async function initMocks() {
   await worker.start();
 }
 
-import black00 from "./assets/stones/black00_128.png";
-import black01 from "./assets/stones/black01_128.png";
-import black02 from "./assets/stones/black02_128.png";
-import black03 from "./assets/stones/black03_128.png";
-import white00 from "./assets/stones/white00_128.png";
-import white01 from "./assets/stones/white01_128.png";
-import white02 from "./assets/stones/white02_128.png";
-import white03 from "./assets/stones/white03_128.png";
-import white04 from "./assets/stones/white04_128.png";
-import white05 from "./assets/stones/white05_128.png";
-import white06 from "./assets/stones/white06_128.png";
-import white07 from "./assets/stones/white07_128.png";
-import white08 from "./assets/stones/white08_128.png";
-import white09 from "./assets/stones/white09_128.png";
-import white10 from "./assets/stones/white10_128.png";
+import {
+  blackImages,
+  whiteImages,
+} from "./utils/imageLoader.js";
+import {
+  updateCanvasSize,
+  drawBoardBackground,
+  drawBoard,
+  getCellSize,
+} from "./utils/boardRenderer.js";
+import { fetchSuggestedMoves } from "./utils/suggestions.js";
 import leftIcon from "./assets/icons/left.svg";
 import rightIcon from "./assets/icons/right.svg";
 import resetIcon from "./assets/icons/reset.svg";
@@ -51,31 +47,6 @@ const boardCanvas = document.createElement("canvas");
 const boardCtx = boardCanvas.getContext("2d");
 const game = new GoGame(9);
 
-const blackSources = [black00, black01, black02, black03];
-const whiteSources = [
-  white00,
-  white01,
-  white02,
-  white03,
-  white04,
-  white05,
-  white06,
-  white07,
-  white08,
-  white09,
-  white10,
-];
-
-function createImages(srcArr) {
-  return srcArr.map((src) => {
-    const img = new Image();
-    img.src = src;
-    return img;
-  });
-}
-
-const blackImages = createImages(blackSources);
-const whiteImages = createImages(whiteSources);
 
 let boardImages = Array.from({ length: game.size }, () =>
   Array(game.size).fill(null)
@@ -88,180 +59,21 @@ let hoverColor = null;
 
 let suggestedMoves = [];
 
-let CELL_SIZE;
-
-function fetchSuggestedMoves() {
-  const boardStr = game.boardToString(game.board);
-  return fetch("/api/next-moves", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ board: boardStr }),
-  })
-    .then((res) => res.json())
-    .then((data) => {
-      if (data && Array.isArray(data.moves)) {
-        suggestedMoves = data.moves;
-        drawBoard();
-      }
-    })
-    .catch((err) => console.error(err));
-}
-
-function updateCanvasSize() {
-  const container = canvas.parentElement;
-  const header = document.querySelector(".banner");
-  const controls = document.querySelector(".controls");
-  const suggestion = document.getElementById("suggestion");
-
-  const paddingTop = parseFloat(getComputedStyle(container).paddingTop) || 0;
-  const paddingBottom =
-    parseFloat(getComputedStyle(container).paddingBottom) || 0;
-
-  const availableHeight =
-    window.innerHeight -
-    header.offsetHeight -
-    controls.offsetHeight -
-    suggestion.offsetHeight -
-    paddingTop -
-    paddingBottom;
-
-  const size = Math.min(container.clientWidth, availableHeight);
-
-  canvas.style.width = `${size}px`;
-  canvas.style.height = `${size}px`;
-  canvas.width = size;
-  canvas.height = size;
-  boardCanvas.width = size;
-  boardCanvas.height = size;
-
-  CELL_SIZE = canvas.width / (game.size + 1);
-}
-
-function drawBoardBackground() {
-  boardCtx.clearRect(0, 0, boardCanvas.width, boardCanvas.height);
-  const extra = 5; // padding around the outer stones
-  const boardStart = CELL_SIZE / 2 - extra;
-  const boardSize = CELL_SIZE * game.size + extra * 2;
-  boardCtx.fillStyle = "#DDB06D";
-  boardCtx.fillRect(boardStart, boardStart, boardSize, boardSize);
-  boardCtx.strokeStyle = "#000";
-  boardCtx.lineWidth = 1;
-  for (let i = 0; i < game.size; i++) {
-    boardCtx.beginPath();
-    boardCtx.moveTo(CELL_SIZE, CELL_SIZE * (i + 1));
-    boardCtx.lineTo(CELL_SIZE * game.size, CELL_SIZE * (i + 1));
-    boardCtx.stroke();
-
-    boardCtx.beginPath();
-    boardCtx.moveTo(CELL_SIZE * (i + 1), CELL_SIZE);
-    boardCtx.lineTo(CELL_SIZE * (i + 1), CELL_SIZE * game.size);
-    boardCtx.stroke();
-  }
-
-  // draw star points for 9x9
-  const star = [
-    [2, 2],
-    [2, 6],
-    [6, 2],
-    [6, 6],
-    [4, 4],
-  ];
-  for (const [x, y] of star) {
-    boardCtx.beginPath();
-    boardCtx.fillStyle = "#000";
-    boardCtx.arc((x + 1) * CELL_SIZE, (y + 1) * CELL_SIZE, 3, 0, Math.PI * 2);
-    boardCtx.fill();
-  }
-}
-
-function drawBoard() {
-  ctx.clearRect(0, 0, canvas.width, canvas.height);
-  ctx.drawImage(boardCanvas, 0, 0);
-  drawStones();
-  drawSuggestedMoves();
-  drawHoverStone();
-}
-
-function drawStones() {
-  const size = CELL_SIZE - 4;
-  for (let x = 0; x < game.size; x++) {
-    for (let y = 0; y < game.size; y++) {
-      if (game.board[x][y] !== 0) {
-        const img = boardImages[x][y];
-        if (img) {
-          ctx.drawImage(
-            img,
-            (x + 1) * CELL_SIZE - size / 2,
-            (y + 1) * CELL_SIZE - size / 2,
-            size,
-            size
-          );
-        } else {
-          ctx.beginPath();
-          ctx.fillStyle = game.board[x][y] === 1 ? "#000" : "#fff";
-          ctx.arc(
-            (x + 1) * CELL_SIZE,
-            (y + 1) * CELL_SIZE,
-            CELL_SIZE / 2 - 2,
-            0,
-            Math.PI * 2
-          );
-          ctx.fill();
-        }
-      }
-    }
-  }
-}
-
-function drawSuggestedMoves() {
-  if (!suggestedMoves.length) return;
-  const radius = CELL_SIZE / 2 - 2;
-  const max = Math.max(...suggestedMoves.map((m) => m.count));
-  const startColor = { r: 173, g: 216, b: 230 };
-  const endColor = { r: 0, g: 0, b: 139 };
-  ctx.textAlign = "center";
-  ctx.textBaseline = "middle";
-  ctx.font = `${CELL_SIZE / 2}px Arial`;
-  for (const { move, count } of suggestedMoves) {
-    const [x, y] = move;
-    if (game.board[x][y] !== 0) continue;
-    const ratio = max === 0 ? 0 : count / max;
-    const r = Math.round(startColor.r + ratio * (endColor.r - startColor.r));
-    const g = Math.round(startColor.g + ratio * (endColor.g - startColor.g));
-    const b = Math.round(startColor.b + ratio * (endColor.b - startColor.b));
-    ctx.fillStyle = `rgb(${r},${g},${b})`;
-    ctx.globalAlpha = 0.6;
-    ctx.beginPath();
-    ctx.arc((x + 1) * CELL_SIZE, (y + 1) * CELL_SIZE, radius, 0, Math.PI * 2);
-    ctx.fill();
-    ctx.globalAlpha = 1;
-    ctx.fillStyle = ratio > 0.5 ? "#fff" : "#000";
-    ctx.fillText(count, (x + 1) * CELL_SIZE, (y + 1) * CELL_SIZE);
-  }
-}
-
-function drawHoverStone() {
-  if (!hoverPos) return;
-  const { x, y } = hoverPos;
-  if (x < 0 || x >= game.size || y < 0 || y >= game.size) return;
-  if (game.board[x][y] !== 0) return;
-  const size = CELL_SIZE - 4;
-  if (!hoverImg) return;
-  ctx.globalAlpha = 0.6;
-  ctx.drawImage(
-    hoverImg,
-    (x + 1) * CELL_SIZE - size / 2,
-    (y + 1) * CELL_SIZE - size / 2,
-    size,
-    size
+function updateSuggestions() {
+  fetchSuggestedMoves(
+    game,
+    (moves) => {
+      suggestedMoves = moves;
+    },
+    () => drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game)
   );
-  ctx.globalAlpha = 1;
 }
 
 canvas.addEventListener("click", (e) => {
   const rect = canvas.getBoundingClientRect();
-  const x = Math.round((e.clientX - rect.left) / CELL_SIZE - 1);
-  const y = Math.round((e.clientY - rect.top) / CELL_SIZE - 1);
+  const cell = getCellSize();
+  const x = Math.round((e.clientX - rect.left) / cell - 1);
+  const y = Math.round((e.clientY - rect.top) / cell - 1);
   const oldBoard = game.cloneBoard(game.board);
   const prevIndex = game.currentIndex;
   if (game.attemptPlace(x, y)) {
@@ -277,15 +89,16 @@ canvas.addEventListener("click", (e) => {
     boardImagesHistory.push(boardImages.map((row) => row.slice()));
     hoverImg = null;
     suggestedMoves = [];
-    drawBoard();
-    fetchSuggestedMoves();
+    drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+    updateSuggestions();
   }
 });
 
 canvas.addEventListener("mousemove", (e) => {
   const rect = canvas.getBoundingClientRect();
-  const x = Math.round((e.clientX - rect.left) / CELL_SIZE - 1);
-  const y = Math.round((e.clientY - rect.top) / CELL_SIZE - 1);
+  const cell = getCellSize();
+  const x = Math.round((e.clientX - rect.left) / cell - 1);
+  const y = Math.round((e.clientY - rect.top) / cell - 1);
   const posChanged = !hoverPos || hoverPos.x !== x || hoverPos.y !== y;
   const colorChanged = hoverColor !== game.currentPlayer;
   if (!posChanged && !colorChanged) {
@@ -308,13 +121,13 @@ canvas.addEventListener("mousemove", (e) => {
   } else {
     hoverImg = null;
   }
-  drawBoard();
+  drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
 });
 
 canvas.addEventListener("mouseleave", () => {
   hoverPos = null;
   hoverImg = null;
-  drawBoard();
+  drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
 });
 
 document.getElementById("back").addEventListener("click", () => {
@@ -323,8 +136,8 @@ document.getElementById("back").addEventListener("click", () => {
       row.slice()
     );
     suggestedMoves = [];
-    drawBoard();
-    fetchSuggestedMoves();
+    drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+    updateSuggestions();
   }
 });
 
@@ -334,8 +147,8 @@ document.getElementById("forward").addEventListener("click", () => {
       row.slice()
     );
     suggestedMoves = [];
-    drawBoard();
-    fetchSuggestedMoves();
+    drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+    updateSuggestions();
   }
 });
 
@@ -347,24 +160,24 @@ document.getElementById("clear").addEventListener("click", () => {
   boardImagesHistory = [boardImages.map((row) => row.slice())];
   hoverImg = null;
   suggestedMoves = [];
-  drawBoard();
-  fetchSuggestedMoves();
+  drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+  updateSuggestions();
 });
 
-window.addEventListener("resize", () => {
-  updateCanvasSize();
-  drawBoardBackground();
-  drawBoard();
-});
+  window.addEventListener("resize", () => {
+    updateCanvasSize(canvas, boardCanvas, game);
+    drawBoardBackground(boardCtx, game);
+    drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+  });
 
 async function init() {
   if (import.meta.env.MODE === "development") {
     await initMocks();
   }
-  updateCanvasSize();
-  drawBoardBackground();
-  drawBoard();
-  fetchSuggestedMoves();
+  updateCanvasSize(canvas, boardCanvas, game);
+  drawBoardBackground(boardCtx, game);
+  drawBoard(ctx, boardCanvas, boardImages, suggestedMoves, hoverPos, hoverImg, game);
+  updateSuggestions();
 }
 
 init();
